quiz_master_22f3002782/backend/requirements.txt
# Core Flask dependencies
Flask==3.1.1
flask-cors==6.0.1
flask-jwt-extended>=4.7.1
flask-security-too>=5.6.2
flask-sqlalchemy>=3.1.1

# Security and password hashing
bcrypt==4.3.0
Werkzeug==3.1.3

# Template engine (comes with Flask but listed for clarity)
Jinja2==3.1.6
MarkupSafe==3.0.2

# Utilities
itsdangerous==2.2.0
blinker==1.9.0
click==8.2.1
celery==5.5.3
redis==5.2.1
